veri-rag/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "veri-rag"
version = "0.1.0"
description = "VERI-RAG: Evidence-Carrying Self-Healing Defense for Secure Retrieval-Augmented Generation"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Nazish Baliyan"},
]
keywords = ["rag", "security", "llm", "retrieval-augmented-generation", "adversarial"]
dependencies = [
    "pydantic>=2.5,<3.0",
    "pyyaml>=6.0",
    "typer>=0.9",
    "rich>=13.7",
    "numpy>=1.26",
    "pandas>=2.1",
    "scikit-learn>=1.3",
    "networkx>=3.2",
    "pytest>=7.4",
    "eval-type-backport>=0.2",
]

[project.scripts]
veri-rag = "veri_rag.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
